+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Added the `suggest_actions` tool to suggest top available actions on the page.
+- Added pre-commit hook, next-feature SKILL.
+
 ## [0.2.0] - 2026-08-23
 
 ### Added